Change the metric endpoint
In the past, metric data was available at the following path:
/metrics/[metric-id]

With this change, metric data is available at the following path
instead:
/experiments/[experiment-id]/metrics/[metric-id]

In the real world, the value of experiment-id would matter. In this mock
API, however, the value of experiment-id has no effect simply because it
was easier to write that way.